requirements.txt
alembic==1.9.1
barnum==0.5.1
bcrypt==4.0.1
blinker==1.5
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-mailman==0.3.0
Flask-Migrate==4.0.1
Flask-Principal==0.4.0
Flask-Security-Too==5.1.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
future==0.18.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
jwt==1.3.1
Mako==1.2.4
MarkupSafe==2.1.1
mkdocs-material-extensions==1.1.1
mysql-connector-python==8.0.31
passlib==1.7.4
protobuf==3.20.1
pycparser==2.21
SQLAlchemy==1.4.46
Werkzeug==2.2.2
WTForms==3.0.1
